@@ -1,17 +1,17 @@
 name: aws_cloudwatch
 description: An AWS CloudWatch wrapper package for easy Flutter cloud logging. Currently only logging is supported.
-version: 0.1.1
+version: 0.1.2
 homepage: https://github.com/Zsmerritt/Flutter_AWS_CloudWatch
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  aws_request: ^0.1.0 # Making aws requests
+  aws_request: ^0.1.1 # Making aws requests
 
 
 dev_dependencies:
